# Conflicts:
#	WzComparerR2.Avatar/Properties/Resources.Designer.cs
#	WzComparerR2.Avatar/UI/AvatarForm.Designer.cs
#	WzComparerR2.Avatar/UI/AvatarForm.cs
#	WzComparerR2.Avatar/WzComparerR2.Avatar.csproj
#	WzComparerR2.Common/Controls/AnimationRecoder.cs
#	WzComparerR2.Common/WzComparerR2.Common.csproj
#	WzComparerR2.MapRender/UI/Tooltip2.cs
#	WzComparerR2.WzLib/Wz_Image.cs
#	WzComparerR2.WzLib/Wz_Png.cs
#	WzComparerR2.WzLib/Wz_Sound.cs
#	WzComparerR2/Comparer/EasyComparer.cs
#	WzComparerR2/MainForm.Designer.cs
#	WzComparerR2/MainForm.cs

+++ WzComparerR2.WzLib/Wz_Image.cs
@@ -152,7 +152,7 @@ namespace WzComparerR2.WzLib
             }
         }
 
-        private void ExtractImg(long offset, Wz_Node parent, int eob)
+        private void ExtractImg(long offset, Wz_Node parent, long eob)
         {
             int entries = 0;
             string tag = this.WzFile.ReadString(offset);
@@ -187,7 +187,7 @@ namespace WzComparerR2.WzLib
                     int form = this.WzFile.ReadInt32() + this.WzFile.BReader.ReadByte();
                     this.WzFile.FileStream.Position += 4;
                     int bufsize = this.WzFile.BReader.ReadInt32();
-                    parent.Value = new Wz_Png(w, h, bufsize - 1, form, (int)this.WzFile.FileStream.Position + 1, this);
+                    parent.Value = new Wz_Png(w, h, bufsize - 1, form, (uint)this.WzFile.FileStream.Position + 1, this);
                     this.WzFile.FileStream.Position += bufsize;
                     break;
 
@@ -203,9 +203,9 @@ namespace WzComparerR2.WzLib
                     this.WzFile.FileStream.Position++;
                     int len = this.WzFile.ReadInt32();
                     int ms = this.WzFile.ReadInt32();
-                    int headerLen = eob - len - (int)this.WzFile.FileStream.Position;
+                    int headerLen = (int)(eob - len - this.WzFile.FileStream.Position);
                     byte[] header = this.WzFile.BReader.ReadBytes(headerLen);
-                    parent.Value = new Wz_Sound(eob - len, len, header, ms, this);
+                    parent.Value = new Wz_Sound((uint)(eob - len), len, header, ms, this);
                     this.WzFile.FileStream.Position = eob;
                     break;
 
@@ -313,7 +313,7 @@ namespace WzComparerR2.WzLib
                     break;
 
                 case 0x09:
-                    ExtractImg(offset, parent, this.WzFile.BReader.ReadInt32() + (int)this.WzFile.FileStream.Position);
+                    ExtractImg(offset, parent, this.WzFile.BReader.ReadInt32() + this.WzFile.FileStream.Position);
                     break;
 
                 default:
